we have a color scheme & fab was renamed
